+++ unity/CAVE/Assets/Resources/CreateScene.cs
@@ -12,6 +12,7 @@ public class CreateScene : MonoBehaviour
         InstantiationResult instantiatedScene = SceneTemplateService.Instantiate(
             Resources.Load<SceneTemplateAsset>("CAVE"),
             false,
+            // TODO: Move to Assets/Resources/Scenes
             "Assets/Scenes/TemplateScene.unity"
         );
         GameObject story = instantiatedScene.scene.GetRootGameObjects()[1];
